PIM-3743: move migrations into a dedicated folder
