+++ app/config/pim_parameters.yml
@@ -34,5 +34,5 @@ parameters:
     # mongodb_server: 'mongodb://localhost:27017'
     # mongodb_database: akeneo_pim
 
-    doctrine_migrations_dir: %kernel.root_dir%/DoctrineMigrations
+    doctrine_migrations_dir: %kernel.root_dir%/../upgrades/schema
