+++ composer.json
@@ -13,6 +13,9 @@
             "": "src/",
             "Context": "features/",
             "OroEmail\\": "app/emails/"
+        },
+        "psr-4": {
+            "Pim\\Upgrade\\": "upgrades/"
         }
     },
     "require": {
