+++ src/Pim/Bundle/EnrichBundle/Resources/config/pim.yml
@@ -49,4 +49,6 @@ services:
               db_id_col:      id
 
 doctrine_migrations:
-    dir_name:             %doctrine_migrations_dir%
+    dir_name: %doctrine_migrations_dir%
+    namespace: Pim\Upgrade\Schema
+    name: PIM Migrations
