+++ upgrades/schema/Version_1_3_20150112174239_remove_media_filepath.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Application\Migrations;
+namespace Pim\Upgrade\Schema;
 
 use Doctrine\DBAL\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
