+++ upgrades/schema/Version_1_3_20150203164205_user_notification.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Application\Migrations;
+namespace Pim\Upgrade\Schema;
 
 use Doctrine\DBAL\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
